Put fast CSE functions inside cse_main.py (WIP)
